refactor(logic): hoist order projections to neutral module (Refs #3290)
Move projectOrderEffects and ProjectedEffects from orders/ to lib/src/projections/
so the turn-layer dry-run no longer creates a forbidden orders->turn edge.
Clear the Phase 0 DAG grandfather allowlist and tighten tests/spec accordingly.
