+++ packages/colonizethis_logic/lib/src/orders/orders.dart
@@ -6,7 +6,7 @@ export 'draft_orders_mutations.dart';
 export 'incremental_candidate_validator.dart';
 export 'order_engine.dart';
 export 'order_merge.dart';
-export 'order_projections.dart';
+export '../projections/order_projections.dart';
 export 'order_suggestion.dart';
 export 'order_suggestion_api.dart';
 export 'order_suggestion_api_impl.dart';
@@ -17,5 +17,5 @@ export 'order_visibility.dart';
 export 'orders_application.dart';
 export 'orders_application_context.dart' show appendMilitaryRegimentToArmy;
 export 'orders_application_helpers.dart';
-export 'projected_effects.dart';
+export '../projections/projected_effects.dart';
 export 'work_order_duration_preview.dart';
