+++ packages/colonizethis_logic/lib/src/projections/projected_effects.dart
