+++ tool/check_logic_domain_import_dag.dart
@@ -24,7 +24,7 @@ const _domains = <String>{
 
 /// Shared modules that may be imported from any domain without counting as a
 /// cross-domain edge (hoisted prerequisites for package split).
-const _neutralTopLevelDirs = <String>{'trace', 'validation'};
+const _neutralTopLevelDirs = <String>{'projections', 'trace', 'validation'};
 
 const _neutralTopLevelFiles = <String>{
   'turn_resolution_seeds.dart',
@@ -55,14 +55,10 @@ const _forbiddenEdges = <(String, String)>{
 /// Grandfathered violations documented in SPEC/program/logic-package-split-phase0.md
 /// until the matching C0 follow-up lands. Format: `fromDomain->toDomain:relative/import/path`.
 const _grandfatherAllowlist = <String>{
-  // `orders -> turn` remains only for the projection dry-run, which still
-  // imports `turn/turn_resolver.dart` (deferred inversion). The trace-runtime
-  // imports in `orders_application*.dart` were already hoisted to the neutral
-  // `lib/src/trace/` module, and the `economy -> orders` projection import was
-  // removed by inverting `tradeOrderValidationContextFromGame` to receive a
-  // caller-computed projected treasury delta (Refs #3290,
-  // SPEC/program/logic-package-split-phase0.md § economy ↔ orders).
-  'orders->turn:orders/order_projections.dart',
+  // Phase 0 C0: all documented wrong-direction edges are eliminated. The
+  // projection dry-run (`projectOrderEffects`) lives in the neutral
+  // `lib/src/projections/` module (like `trace/` and `validation/`), so
+  // `orders` no longer imports `turn/` directly (Refs #3290).
 };
 
 void main() {
